requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
jwt==1.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
aioredis==2.0.1
redis==5.0.1
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3
apscheduler==3.10.4
httpx==0.25.2
python-multipart==0.0.6
cors==1.0.1
slowapi==0.1.9
